Webshop v1
This is a demo of armoryd which is done via a webshop.
Webshop is fully functional and by default runs on testnet.
You can specify when running, whether to accept bitcoins into a normal armory wallet or a custom lockbox. Either way, a new address is generated for each order and automatically detects when payment is made.